pyproject.toml
[project]
name = "anthony-portfolio-backend"
version = "0.1.0"
description = "Portfolio backend API"
authors = [{ name = "Anthony Sung" }]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.30.0,<1.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
    "pydantic[email-validation] (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "aiosmtplib (>=5.1.0,<6.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
